Convert codebase to modules, use class instead of function
